refactor(daemon): split orchestrator agent loop
